feat(Status): Restructure PTeamServiceDetails components into a dedicated directory for improved organization
